Релизная версия. переписан код опроса устройств. убраны лаги в меню.
